requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.33.0
azure-mgmt-core==1.5.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
fastapi-mcp==0.3.0
griffe==1.7.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
jiter==0.9.0
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
msrest==0.7.1
oauthlib==3.2.2
openai==1.75.0
openai-agents==0.0.12
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
unicorn==2.1.3
urllib3==2.4.0
uvicorn==0.34.2
